requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
jinja2>=3.1.0
httpx>=0.27.0
python-telegram-bot[job-queue]>=21.0
python-dotenv>=1.0.0
